Move test utilities to :core:testing module
