+++ feature/stopdetail/build.gradle
@@ -54,4 +54,5 @@ dependencies {
 
     implementation project(":core:data")
     implementation project(":core:model")
+    implementation project(":core:testing")
 }
