+++ feature/stoplist/build.gradle
@@ -55,4 +55,5 @@ dependencies {
     implementation project(":core:data")
     implementation project(":core:model")
     implementation project(":core:ui")
+    implementation project(":core:testing")
 }
